+++ Wifi/WifiInterface.h
@@ -5,79 +5,7 @@
 #include <shared_mutex>
 #include <unordered_map>
 #include "Signal.h"
-
-
-struct WifiDevice {
-    enum class State {
-        UnknownState,
-        UnpairedState,
-        WaitingAuthenState,
-        AuthenFailState,
-        AuthenSuccessState,
-        PairedState,
-        PairingState,
-        PairingFailState,
-        ConnectedState
-    };
-
-    enum class DeviceType {
-        Unknown,
-        Connected,
-        Paired,
-        Unpaired
-    };
-
-    enum class SpeedMode {
-        Weak,
-        Medium,
-        Strong
-    };
-
-    enum class IP {
-        Automatic,
-        Manual,
-        BootP
-    };
-
-    enum class DNS {
-        Automatic,
-        Manual
-    };
-
-    enum class Proxy {
-        Off,
-        Manual,
-        Automatic
-    };
-
-    std::string mName, mPassword, mAddress, mIPAddress, mSubnet, mRouter;
-    bool mPrivateAddr, mLowDataMode, mLimitIP;
-    SpeedMode mSpeedMode;
-    IP mIP;
-    DeviceType mDeviceType;
-    DNS mDNS;
-    Proxy mProxy;
-    State mState;
-
-    WifiDevice(std::string name, std::string addr, DeviceType type = DeviceType::Unknown, SpeedMode speedMode = SpeedMode::Strong, bool privateAddr = true, IP ip = IP::Automatic,
-               DNS dns = DNS::Automatic, Proxy proxy = Proxy::Automatic, std::string password = "", bool lowDataMode = true, bool limitIP = false,
-               std::string ipAddr = "", std::string subnetMask = "", std::string router = "") {
-        mName = name;
-        mAddress = addr;
-        mSpeedMode = speedMode;
-        mDeviceType = type;
-        mPrivateAddr = privateAddr;
-        mIP = ip;
-        mDNS = dns;
-        mProxy = proxy;
-        mPassword = password;
-        mLowDataMode = lowDataMode;
-        mLimitIP = limitIP;
-        mIPAddress = ipAddr;
-        mSubnet = subnetMask;
-        mRouter = router;
-    }
-};
+#include "WifiDevice.h"
 
 class WifiIFConnect;
 class WifiAdapter;
